Sentence case for <title> section headers.
There was inconsistency in the documentation, especially between older
and newer documenation. The older documenation used title case, but the
newer files used a combination of title and sentence case.

Every page on www.g.o and most-every page on wiki.g.o now uses sentence
case so I am pushing for it across the board for Gentoo's documentation.

Signed-off-by: Matthew Marchese <maffblaster@gentoo.org>